split configuration variables and functions
in order to allow another program to reuse these functions (for
example ansibullbot) without defining the ansible variables.
