+++ lib/ansible/constants.py
@@ -20,126 +20,14 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 import os
-import tempfile
 from string import ascii_letters, digits
 
+from ansible.config import get_config, mk_boolean
 from ansible.errors import AnsibleOptionsError
-from ansible.module_utils.six import string_types
-from ansible.module_utils.six.moves import configparser
 from ansible.module_utils._text import to_text
-from ansible.parsing.quoting import unquote
-from ansible.utils.path import makedirs_safe
+from ansible.module_utils.six.moves import configparser
 
-BOOL_TRUE = frozenset( [ "true", "t", "y", "1", "yes", "on" ] )
-
-
-def mk_boolean(value):
-    ret = value
-    if not isinstance(value, bool):
-        if value is None:
-            ret = False
-        ret = (str(value).lower() in BOOL_TRUE)
-    return ret
-
-
-def shell_expand(path, expand_relative_paths=False):
-    '''
-    shell_expand is needed as os.path.expanduser does not work
-    when path is None, which is the default for ANSIBLE_PRIVATE_KEY_FILE
-    '''
-    if path:
-        path = os.path.expanduser(os.path.expandvars(path))
-        if expand_relative_paths and not path.startswith('/'):
-            # paths are always 'relative' to the config?
-            if 'CONFIG_FILE' in globals():
-                CFGDIR = os.path.dirname(CONFIG_FILE)
-                path = os.path.join(CFGDIR, path)
-            path = os.path.abspath(path)
-    return path
-
-
-def get_config(p, section, key, env_var, default, value_type=None, expand_relative_paths=False):
-    ''' return a configuration variable with casting
-
-    :arg p: A ConfigParser object to look for the configuration in
-    :arg section: A section of the ini config that should be examined for this section.
-    :arg key: The config key to get this config from
-    :arg env_var: An Environment variable to check for the config var.  If
-        this is set to None then no environment variable will be used.
-    :arg default: A default value to assign to the config var if nothing else sets it.
-    :kwarg value_type: The type of the value.  This can be any of the following strings:
-        :boolean: sets the value to a True or False value
-        :integer: Sets the value to an integer or raises a ValueType error
-        :float: Sets the value to a float or raises a ValueType error
-        :list: Treats the value as a comma separated list.  Split the value
-            and return it as a python list.
-        :none: Sets the value to None
-        :path: Expands any environment variables and tilde's in the value.
-        :tmp_path: Create a unique temporary directory inside of the directory
-            specified by value and return its path.
-        :pathlist: Treat the value as a typical PATH string.  (On POSIX, this
-            means colon separated strings.)  Split the value and then expand
-            each part for environment variables and tildes.
-    :kwarg expand_relative_paths: for pathlist and path types, if this is set
-        to True then also change any relative paths into absolute paths.  The
-        default is False.
-    '''
-    value = _get_config(p, section, key, env_var, default)
-    if value_type == 'boolean':
-        value = mk_boolean(value)
-
-    elif value:
-        if value_type == 'integer':
-            value = int(value)
-
-        elif value_type == 'float':
-            value = float(value)
-
-        elif value_type == 'list':
-            if isinstance(value, string_types):
-                value = [x.strip() for x in value.split(',')]
-
-        elif value_type == 'none':
-            if value == "None":
-                value = None
-
-        elif value_type == 'path':
-            value = shell_expand(value, expand_relative_paths=expand_relative_paths)
-
-        elif value_type == 'tmppath':
-            value = shell_expand(value)
-            if not os.path.exists(value):
-                makedirs_safe(value, 0o700)
-            prefix = 'ansible-local-%s' % os.getpid()
-            value = tempfile.mkdtemp(prefix=prefix, dir=value)
-
-        elif value_type == 'pathlist':
-            if isinstance(value, string_types):
-                value = [shell_expand(x, expand_relative_paths=expand_relative_paths) for x in value.split(os.pathsep)]
-
-        elif isinstance(value, string_types):
-            value = unquote(value)
-
-    return to_text(value, errors='surrogate_or_strict', nonstring='passthru')
-
-
-def _get_config(p, section, key, env_var, default):
-    ''' helper function for get_config '''
-    value = default
-
-    if p is not None:
-        try:
-            value = p.get(section, key, raw=True)
-        except:
-            pass
-
-    if env_var is not None:
-        env_value = os.environ.get(env_var, None)
-        if env_value is not None:
-            value = env_value
-
-    return to_text(value, errors='surrogate_or_strict', nonstring='passthru')
-
+mk_boolean = mk_boolean # TODO
 
 def load_config_file():
     ''' Load Config File order(first found is used): ENV, CWD, HOME, /etc/ansible '''
@@ -169,6 +57,7 @@ def load_config_file():
 
 
 p, CONFIG_FILE = load_config_file()
+CFGDIR = os.path.dirname(CONFIG_FILE)
 
 # non configurable but used as defaults
 BLACKLIST_EXTS = ('.pyc', '.pyo', '.swp', '.bak', '~', '.rpm', '.md', '.txt')
