+++ lib/ansible/executor/playbook_executor.py
@@ -192,7 +192,7 @@ class PlaybookExecutor:
                         retries = sorted(retries)
                         if len(retries) > 0:
                             if C.RETRY_FILES_SAVE_PATH:
-                                basedir = C.shell_expand(C.RETRY_FILES_SAVE_PATH)
+                                basedir = C.shell_expand(C.RETRY_FILES_SAVE_PATH, C.CFGDIR)
                             elif playbook_path:
                                 basedir = os.path.dirname(os.path.abspath(playbook_path))
                             else:
